+++ src/commandObject.js
@@ -1,4 +1,4 @@
-exports.CommandObject = {
+module.exports = {
     command: String,
     params: Object,
     response: Object,